+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Development
 
+* Use RuboCop's `--force-exclusion` option to always ignore files specified in the `Exclude` configuration in `.rubocop.yml`
+
 ## v1.0.2
 
 * Support both spelling “offense” (RuboCop 0.19 or later) and “offence” (prior to RuboCop 0.19) ([rubocop#700](https://github.com/bbatsov/rubocop/issues/700))